consolidate "init" modules in their own folder
