+++ src/dfp/fill-static-advert-slots.ts
@@ -3,12 +3,12 @@ import type { SizeMapping } from 'core/ad-sizes';
 import { adSizes, createAdSize } from 'core/ad-sizes';
 import { getCurrentBreakpoint } from 'detect/detect-breakpoint';
 import { commercialFeatures } from 'lib/commercial-features';
-import { removeDisabledSlots } from '../lib/remove-slots';
+import { setupPrebidOnce } from '../init/consented/prepare-prebid';
+import { removeDisabledSlots } from '../init/consented/remove-slots';
 import { createAdvert } from './create-advert';
 import { dfpEnv } from './dfp-env';
 import { displayAds } from './display-ads';
 import { displayLazyAds } from './display-lazy-ads';
-import { setupPrebidOnce } from './prepare-prebid';
 import { queueAdvert } from './queue-advert';
 
 const decideAdditionalSizes = (adSlot: HTMLElement): SizeMapping => {
