+++ src/init/consented/prepare-googletag.ts
@@ -12,25 +12,25 @@ import { commercialFeatures } from 'lib/commercial-features';
 import raven from 'lib/raven';
 import { init as initMeasureAdLoad } from 'messenger/measure-ad-load';
 import { reportError } from 'utils/report-error';
-import { removeSlots } from '../lib/remove-slots';
-import { init as background } from '../messenger/background';
-import { init as sendClick } from '../messenger/click';
-import { init as disableRefresh } from '../messenger/disable-refresh';
-import { init as fullwidth } from '../messenger/full-width';
-import { init as initGetPageTargeting } from '../messenger/get-page-targeting';
-import { init as initGetPageUrl } from '../messenger/get-page-url';
-import { init as getStyles } from '../messenger/get-stylesheet';
-import { init as passback } from '../messenger/passback';
-import { init as passbackRefresh } from '../messenger/passback-refresh';
-import { init as resize } from '../messenger/resize';
-import { init as scroll } from '../messenger/scroll';
-import { init as type } from '../messenger/type';
-import { init as viewport } from '../messenger/viewport';
-import { createSlotFillListener } from './fill-slot-listener';
-import { fillStaticAdvertSlots } from './fill-static-advert-slots';
-import { onSlotLoad } from './on-slot-load';
-import { onSlotRender } from './on-slot-render';
-import { onSlotViewableFunction } from './on-slot-viewable';
+import { createSlotFillListener } from '../../dfp/fill-slot-listener';
+import { fillStaticAdvertSlots } from '../../dfp/fill-static-advert-slots';
+import { onSlotLoad } from '../../dfp/on-slot-load';
+import { onSlotRender } from '../../dfp/on-slot-render';
+import { onSlotViewableFunction } from '../../dfp/on-slot-viewable';
+import { init as background } from '../../messenger/background';
+import { init as sendClick } from '../../messenger/click';
+import { init as disableRefresh } from '../../messenger/disable-refresh';
+import { init as fullwidth } from '../../messenger/full-width';
+import { init as initGetPageTargeting } from '../../messenger/get-page-targeting';
+import { init as initGetPageUrl } from '../../messenger/get-page-url';
+import { init as getStyles } from '../../messenger/get-stylesheet';
+import { init as passback } from '../../messenger/passback';
+import { init as passbackRefresh } from '../../messenger/passback-refresh';
+import { init as resize } from '../../messenger/resize';
+import { init as scroll } from '../../messenger/scroll';
+import { init as type } from '../../messenger/type';
+import { init as viewport } from '../../messenger/viewport';
+import { removeSlots } from './remove-slots';
 
 initMessenger(
 	[
