+++ src/init/consented/prepare-prebid.spec.ts
@@ -7,7 +7,7 @@ import type { TCFv2ConsentState } from '@guardian/consent-management-platform/di
 import { log } from '@guardian/libs';
 import { commercialFeatures } from 'lib/commercial-features';
 import { isInCanada } from 'utils/geo-utils';
-import { prebid } from '../header-bidding/prebid/prebid';
+import { prebid } from '../../header-bidding/prebid/prebid';
 import { _ } from './prepare-prebid';
 
 const { setupPrebid } = _;
@@ -25,13 +25,13 @@ jest.mock('lib/commercial-features', () => ({
 	commercialFeatures: {},
 }));
 
-jest.mock('../header-bidding/prebid/prebid', () => ({
+jest.mock('header-bidding/prebid/prebid', () => ({
 	prebid: {
 		initialise: jest.fn(),
 	},
 }));
 
-jest.mock('./Advert', () =>
+jest.mock('dfp/Advert', () =>
 	jest.fn().mockImplementation(() => ({ advert: jest.fn() })),
 );
 
@@ -39,11 +39,11 @@ jest.mock('lib/build-page-targeting', () => ({
 	getPageTargeting: jest.fn(),
 }));
 
-jest.mock('../header-bidding/prebid/bid-config', () => ({
+jest.mock('header-bidding/prebid/bid-config', () => ({
 	isInVariant: jest.fn(),
 }));
 
-jest.mock('../header-bidding/utils', () => ({
+jest.mock('header-bidding/utils', () => ({
 	shouldIncludeOnlyA9: false,
 }));
