+++ src/init/consented/third-party-tags.ts
@@ -11,9 +11,9 @@ import { remarketing } from 'core/third-party-tags/remarketing';
 import { twitter } from 'core/third-party-tags/twitter-uwt';
 import { commercialFeatures } from 'lib/commercial-features';
 import type { ThirdPartyTag } from 'types/global';
-import { imrWorldwide } from '../third-party-tags/imr-worldwide';
-import { imrWorldwideLegacy } from '../third-party-tags/imr-worldwide-legacy';
-import fastdom from './fastdom-promise';
+import fastdom from '../../lib/fastdom-promise';
+import { imrWorldwide } from '../../third-party-tags/imr-worldwide';
+import { imrWorldwideLegacy } from '../../third-party-tags/imr-worldwide-legacy';
 
 const addScripts = (tags: ThirdPartyTag[]) => {
 	const ref = document.scripts[0];
