+++ src/init/consentless/dynamic/article-inline.ts
