+++ src/init/consentless/init-fixed-slots.ts
