+++ src/standalone.commercial.ts
@@ -20,31 +20,31 @@ import { log } from '@guardian/libs';
 import { init as prepareAdVerification } from 'ad-verification/prepare-ad-verification';
 import { EventTimer } from 'core/event-timer';
 import { adSlotIdPrefix } from 'dfp/dfp-env-globals';
-import { init as prepareA9 } from 'dfp/prepare-a9';
-import { init as prepareGoogletag } from 'dfp/prepare-googletag';
-import { initPermutive } from 'dfp/prepare-permutive';
-import { init as preparePrebid } from 'dfp/prepare-prebid';
-import { init as initRedplanet } from 'dfp/redplanet';
-import { adFreeSlotRemove } from 'lib/ad-free-slot-remove';
-import { initAdblockAsk } from 'lib/adblock-ask';
+import { adFreeSlotRemove } from 'init/consented/ad-free-slot-remove';
+import { initAdblockAsk } from 'init/consented/adblock-ask';
+import { init as initArticleAsideAdverts } from 'init/consented/article-aside-adverts';
+import { init as initArticleBodyAdverts } from 'init/consented/article-body-adverts';
+import { initCommentAdverts } from 'init/consented/comment-adverts';
+import { initCommentsExpandedAdverts } from 'init/consented/comments-expanded-advert';
+import { init as initComscore } from 'init/consented/comscore';
+import { init as initHighMerch } from 'init/consented/high-merch';
+import { init as initIpsosMori } from 'init/consented/ipsos-mori';
+import { init as initLiveblogAdverts } from 'init/consented/liveblog-adverts';
+import { init as initMobileSticky } from 'init/consented/mobile-sticky';
+import { init as prepareA9 } from 'init/consented/prepare-a9';
+import { init as prepareGoogletag } from 'init/consented/prepare-googletag';
+import { initPermutive } from 'init/consented/prepare-permutive';
+import { init as preparePrebid } from 'init/consented/prepare-prebid';
+import { init as initRedplanet } from 'init/consented/redplanet';
+import { removeConsentedAdsOnConsentChange } from 'init/consented/remove-consented-ads-on-consent-change';
+import { removeDisabledSlots as closeDisabledSlots } from 'init/consented/remove-slots';
+import { initTeadsCookieless } from 'init/consented/teads-cookieless';
+import { init as initThirdPartyTags } from 'init/consented/third-party-tags';
+import { init as initTrackGpcSignal } from 'init/consented/track-gpc-signal';
+import { init as initTrackScrollDepth } from 'init/consented/track-scroll-depth';
+import { init as setAdTestCookie } from 'init/shared/set-adtest-cookie';
+import { init as setAdTestInLabelsCookie } from 'init/shared/set-adtest-in-labels-cookie';
 import { commercialFeatures } from 'lib/commercial-features';
-import { init as initComscore } from 'lib/comscore';
-import { init as initHighMerch } from 'lib/high-merch';
-import { init as initIpsosMori } from 'lib/ipsos-mori';
-import { init as initMobileSticky } from 'lib/mobile-sticky';
-import { removeConsentedAdsOnConsentChange } from 'lib/remove-consented-ads-on-consent-change';
-import { removeDisabledSlots as closeDisabledSlots } from 'lib/remove-slots';
-import { init as setAdTestCookie } from 'lib/set-adtest-cookie';
-import { init as setAdTestInLabelsCookie } from 'lib/set-adtest-in-labels-cookie';
-import { initTeadsCookieless } from 'lib/teads-cookieless';
-import { init as initThirdPartyTags } from 'lib/third-party-tags';
-import { init as initTrackGpcSignal } from 'lib/track-gpc-signal';
-import { init as initTrackScrollDepth } from 'lib/track-scroll-depth';
-import { init as initArticleAsideAdverts } from 'spacefinder/article-aside-adverts';
-import { init as initArticleBodyAdverts } from 'spacefinder/article-body-adverts';
-import { initCommentAdverts } from 'spacefinder/comment-adverts';
-import { initCommentsExpandedAdverts } from 'spacefinder/comments-expanded-advert';
-import { init as initLiveblogAdverts } from 'spacefinder/liveblog-adverts';
 import { reportError } from 'utils/report-error';
 import { catchErrorsWithContext } from 'utils/robust';
